requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
passlib[bcrypt]==1.7.4
python-jose==3.5.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
email-validator==2.3.0
tensorflow-cpu==2.17.0
google-generativeai==0.7.2
numpy==1.26.4
